git-svn-id: svn://svn.code.sf.net/p/win32-framework/code@2851 40094c20-7226-0410-9e7a-ca24b3fb783e

+++ include/wxx_appcore.h
@@ -892,7 +892,7 @@ namespace Win32xx
             reinterpret_cast<LPWSTR>(&pString), 0);
 
         if (charCount > 0)
-            m_str.assign(pString, charCount);
+            m_str.assign(pString, static_cast<size_t>(charCount));
 
         return (charCount != 0);
     }
